[DXP-1399] PR review fixes
