+++ core/src/test/java/dev/streamx/cli/command/cloud/KubernetesClientProfile.java
@@ -1,4 +1,4 @@
-package dev.streamx.cli.command.cloud.deploy;
+package dev.streamx.cli.command.cloud;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 import java.util.Map;
